K-9 navBar Component creation/implementation
