+++ client/src/components/app.js
@@ -18,4 +18,3 @@ class App extends Component{
 }
 
 export default App;
-
