+++ client/src/stylesheets/css/components/members.css
@@ -23,6 +23,7 @@
     color: gray; }
 
 .mp-container {
+  overflow: hidden;
   display: flex;
   flex-direction: column;
   align-items: center;
