+++ client/src/stylesheets/css/components/navbar.css
@@ -1,0 +1,118 @@
+.body {
+  font-family: Lato, sans-serif; }
+
+.list-no-styles li {
+  text-decoration: none; }
+
+.no-padding {
+  padding: 0; }
+
+.no-margin {
+  margin: 0; }
+
+.text-left {
+  text-align: left; }
+
+.text-center {
+  text-align: center; }
+
+.text-right {
+  text-align: right; }
+
+.page-mid {
+  position: absolute;
+  top: 45%;
+  left: 50%;
+  -webkit-transform: translate(-50%, -50%);
+  -moz-transform: translate(-50%, -50%);
+  -ms-transform: translate(-50%, -50%);
+  -o-transform: translate(-50%, -50%);
+  transform: translate(-50%, -50%); }
+
+.kapitolRed {
+  background: #E73347; }
+
+.no-border {
+  border: none; }
+
+.border-light-gray {
+  border: 2px solid #DFE0E2; }
+
+select {
+  z-index: 0; }
+
+select.dropdown {
+  cursor: pointer;
+  z-index: inherit;
+  margin: 0px;
+  margin-top: 30px;
+  height: 40px;
+  width: 150px;
+  border-width: 1px;
+  border-style: solid;
+  border-color: #DFE0E2;
+  padding: 8px;
+  padding-left: 20px;
+  font-size: 16px;
+  color: #666666;
+  -webkit-appearance: none;
+  -webkit-border-radius: 0px;
+  -moz-appearance: none;
+  appearance: none;
+  background: url("../../images/dropdown.png") no-repeat 85% 55% #ffffff;
+  background-size: 11px 7px; }
+  select.dropdown:hover {
+    outline: none; }
+
+.navigation-bar {
+  width: 100%;
+  border-bottom: 1px solid #DFE0E2; }
+  .navigation-bar img {
+    width: 150px;
+    margin: 35px 0 30px 50px; }
+  .navigation-bar .nav-search-container {
+    float: right;
+    padding: 0 40px; }
+    .navigation-bar .nav-search-container #search-bar {
+      border: 1px solid #DFE0E2;
+      display: inline-block;
+      height: 40px;
+      width: 250px;
+      margin-right: 15px;
+      padding: 5px; }
+      .navigation-bar .nav-search-container #search-bar:focus {
+        outline: none; }
+    .navigation-bar .nav-search-container #search-box {
+      width: 250px !important;
+      position: absolute;
+      top: 70px; }
+    .navigation-bar .nav-search-container #fp-selector {
+      display: inline-block;
+      height: 50px; }
+
+@media screen and (max-width: 700px) {
+  .navigation-bar {
+    overflow-x: hidden; }
+    .navigation-bar img {
+      width: 150px;
+      display: block;
+      margin: 20px auto;
+      padding: 0; }
+    .navigation-bar .nav-search-container {
+      float: none; }
+      .navigation-bar .nav-search-container #search-bar {
+        display: block;
+        margin: 10px auto;
+        width: 200px; }
+      .navigation-bar .nav-search-container #search-box {
+        position: relative;
+        top: -11px;
+        width: 200px !important; }
+      .navigation-bar .nav-search-container #fp-selector {
+        display: block;
+        margin: 10px auto; }
+        .navigation-bar .nav-search-container #fp-selector select {
+          display: block;
+          margin: 20px auto 30px; } }
+
+/*# sourceMappingURL=navbar.css.map */
