+++ client/src/stylesheets/scss/components/App.scss
@@ -7,11 +7,10 @@
   text-align: center;
   width: 100%;
   #fp-search{
-    //border: 2px solid black;
     margin-top: 35px;
     height: 50px;
-    input{
-      height: 100%;
+    #search-bar{
+      height: 50px;
       width: 350px;
       font-size: 16px;
       padding-left: 10px;
