+++ client/src/stylesheets/scss/components/members.scss
@@ -34,6 +34,7 @@ $kapitolBlue: #007DBC;
 }
 
 .mp-container {
+  overflow: hidden;
   display: flex;
   flex-direction: column;
   align-items: center;
