+++ client/src/stylesheets/scss/components/navbar.scss
@@ -1,0 +1,68 @@
+@import './common';
+
+.navigation-bar{
+  width: 100%;
+  border-bottom: 1px solid $kapitolLightGray;
+  img{
+    width: 150px;
+    margin: 35px 0 30px 50px;
+  }
+  .nav-search-container{
+    float: right;
+    padding: 0 40px;
+    #search-bar{
+      border: 1px solid $kapitolLightGray;
+      display: inline-block;
+      height: 40px;
+      width: 250px;
+      margin-right: 15px;
+      padding: 5px;
+      &:focus{
+        outline: none;
+      }
+    }
+    #search-box{
+      width: 250px !important;
+      position: absolute;
+      top: 70px;
+    }
+    #fp-selector{
+      display: inline-block;
+      height: 50px;
+    }
+  }
+}
+
+@media screen and (max-width:700px){
+  .navigation-bar{
+    overflow-x: hidden;
+    img{
+      width: 150px;
+      display: block;
+      margin: 20px auto;
+      padding: 0;
+    }
+    .nav-search-container{
+      float: none;
+      #search-bar{
+        display: block;
+        margin: 10px auto;
+        width: 200px;
+      }
+      #search-box{
+        position: relative;
+        top: -11px;
+        width: 200px !important;
+      }
+      #fp-selector{
+        display: block;
+        margin: 10px auto;
+        select{
+          display: block;
+          margin: 20px auto 30px;
+        }
+      }
+    }
+  }
+}
+
