+++ client/src/stylesheets/scss/index.scss
@@ -2,3 +2,4 @@
 @import './components/common';
 @import './components/searchbox';
 @import './components/members';
+@import './components/navbar';
